move the outer reader into its own module
