+++ dnp3/src/entry/master/tcp/mod.rs
@@ -3,7 +3,8 @@ use crate::master::error::Shutdown;
 use crate::master::handle::{Listener, MasterConfiguration, MasterHandle};
 use crate::master::session::{MasterSession, RunError};
 use crate::tokio::net::TcpStream;
-use crate::transport::{TransportReader, TransportWriter};
+use crate::transport::reader::TransportReader;
+use crate::transport::TransportWriter;
 use std::future::Future;
 use std::net::SocketAddr;
 use std::time::Duration;
