+++ dnp3/src/transport/mod.rs
@@ -1,30 +1,22 @@
-use crate::app::parse::parser::{ParsedFragment, Request, Response};
-use crate::app::parse::DecodeLogLevel;
 use crate::app::EndpointType;
 use crate::entry::EndpointAddress;
-use crate::link::error::LinkError;
 use crate::link::header::BroadcastConfirmMode;
 use crate::master::session::MasterSession;
 use crate::outstation::task::OutstationSession;
 use crate::outstation::SelfAddressSupport;
-use crate::tokio::io::{AsyncRead, AsyncWrite};
+use crate::transport::reader::TransportReader;
 
 #[cfg(test)]
 pub(crate) mod mock;
 #[cfg(not(test))]
 pub(crate) mod real;
 
-#[cfg(not(test))]
-/// This type definition is used so that we can mock the transport reader during testing.
-/// If Rust eventually allows `async fn` in traits, this can be removed
-pub(crate) type ReaderType = real::reader::Reader;
+pub(crate) mod reader;
+
 #[cfg(not(test))]
 /// This type definition is used so that we can mock the transport writer during testing.
 /// If Rust eventually allows `async fn` in traits, this can be removed
 pub(crate) type TransportWriter = real::writer::Writer;
-
-#[cfg(test)]
-pub(crate) type ReaderType = crate::transport::mock::reader::MockReader;
 #[cfg(test)]
 pub(crate) type TransportWriter = crate::transport::mock::writer::MockWriter;
 
@@ -53,133 +45,6 @@ impl FragmentInfo {
 pub(crate) struct Fragment<'a> {
     pub(crate) info: FragmentInfo,
     pub(crate) data: &'a [u8],
-}
-
-pub(crate) mod constants {
-    pub(crate) const FIN_MASK: u8 = 0b1000_0000;
-    pub(crate) const FIR_MASK: u8 = 0b0100_0000;
-}
-
-pub(crate) struct TransportReader {
-    logged: bool,
-    inner: ReaderType,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub(crate) enum Timeout {
-    Yes,
-    No,
-}
-
-impl TransportReader {
-    fn master(address: EndpointAddress, rx_buffer_size: usize) -> Self {
-        Self {
-            logged: false,
-            inner: ReaderType::master(address, rx_buffer_size),
-        }
-    }
-
-    fn outstation(
-        address: EndpointAddress,
-        self_address_support: SelfAddressSupport,
-        rx_buffer_size: usize,
-    ) -> Self {
-        Self {
-            logged: false,
-            inner: ReaderType::outstation(address, self_address_support, rx_buffer_size),
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn get_inner(&mut self) -> &mut ReaderType {
-        &mut self.inner
-    }
-
-    pub(crate) async fn read_next<T>(&mut self, io: &mut T) -> Result<(), LinkError>
-    where
-        T: AsyncRead + AsyncWrite + Unpin,
-    {
-        self.logged = false;
-        self.inner.read_next(io).await
-    }
-
-    pub(crate) async fn read_with_timeout<T>(
-        &mut self,
-        io: &mut T,
-        deadline: crate::tokio::time::Instant,
-    ) -> Result<Timeout, LinkError>
-    where
-        T: AsyncRead + AsyncWrite + Unpin,
-    {
-        crate::tokio::select! {
-            res = self.read_next(io) => {
-                res?;
-                Ok(Timeout::No)
-            },
-            _ = crate::tokio::time::delay_until(deadline) => {
-                Ok(Timeout::Yes)
-            }
-        }
-    }
-
-    pub(crate) fn reset(&mut self) {
-        self.inner.reset()
-    }
-
-    pub(crate) fn get_response(
-        &mut self,
-        level: DecodeLogLevel,
-    ) -> Option<(EndpointAddress, Response)> {
-        let (log, info, parsed) = self.peek(level)?;
-        match parsed.to_response() {
-            Err(err) => {
-                if log {
-                    log::error!("response error: {}", err)
-                }
-                None
-            }
-            Ok(response) => Some((info.source, response)),
-        }
-    }
-
-    pub(crate) fn get_request(&mut self, level: DecodeLogLevel) -> Option<(FragmentInfo, Request)> {
-        let (log, info, parsed) = self.peek(level)?;
-        match parsed.to_request() {
-            Err(err) => {
-                if log {
-                    log::error!("request error: {}", err)
-                }
-                None
-            }
-            Ok(request) => Some((info, request)),
-        }
-    }
-
-    fn peek(&mut self, level: DecodeLogLevel) -> Option<(bool, FragmentInfo, ParsedFragment)> {
-        let log_this_peek = !self.logged;
-        self.logged = true;
-        let fragment: Fragment = self.inner.peek()?;
-        let level = if log_this_peek {
-            level
-        } else {
-            DecodeLogLevel::Nothing
-        };
-        let parsed = match ParsedFragment::parse(level.receive(), fragment.data) {
-            Err(err) => {
-                if log_this_peek {
-                    log::warn!("error parsing fragment header: {}", err)
-                }
-                return None;
-            }
-            Ok(parsed) => parsed,
-        };
-        if let Err(err) = parsed.objects {
-            if log_this_peek {
-                log::warn!("error parsing object headers: {}", err)
-            }
-        }
-        Some((log_this_peek, fragment.info, parsed))
-    }
 }
 
 pub(crate) fn create_master_transport_layer(
