+++ dnp3/src/transport/real/mod.rs
@@ -3,3 +3,8 @@ pub(crate) mod header;
 pub(crate) mod reader;
 pub(crate) mod sequence;
 pub(crate) mod writer;
+
+pub(crate) mod constants {
+    pub(crate) const FIN_MASK: u8 = 0b1000_0000;
+    pub(crate) const FIR_MASK: u8 = 0b0100_0000;
+}
